+++ packages/core/src/app/address/googleAutocomplete/AddressSelectorAU.test.ts
@@ -1,0 +1,16 @@
+import AddressSelectorAU from "./AddressSelectorAU";
+import { getGoogleAutocompletePlaceMock } from "./googleAutocompleteResult.mock";
+
+describe('AddressSelectorAU', () => {
+    it('should return correct street address', () => {
+        const selector = new AddressSelectorAU(getGoogleAutocompletePlaceMock());
+
+        expect(selector.getStreet()).toBe('1-3 (l) Smail Street');
+    });
+
+    it('should return correct street2 address value', () => {
+        const selector = new AddressSelectorAU(getGoogleAutocompletePlaceMock());
+
+        expect(selector.getStreet2()).toBe('unit 6');
+    });
+});