+++ packages/stripe-integration/src/stripe-ocs/StripeOCSPaymentMethod.tsx
@@ -26,16 +26,17 @@ const StripeOCSPaymentMethod: FunctionComponent<PaymentMethodProps> = ({
 }) => {
     const collapseStripeElement = useRef<() => void>();
     const { onToggle, selectedItemId } = useContext(AccordionContext);
-    const containerId = `stripe-${method.id}-component-field`;
+    const methodSelector = `${method.gateway}-${method.id}`;
+    const containerId = `${methodSelector}-component-field`;
     const paymentContext = paymentForm;
 
     useEffect(() => {
-        if (selectedItemId?.includes('stripeupe-')) {
+        if (selectedItemId?.includes(`${method.gateway}-`)) {
             return;
         }
 
         collapseStripeElement.current?.();
-    }, [selectedItemId]);
+    }, [selectedItemId, method.gateway]);
 
     const renderSubmitButton = useCallback(() => {
         paymentContext.hidePaymentSubmitButton(method, false);
@@ -60,8 +61,15 @@ const StripeOCSPaymentMethod: FunctionComponent<PaymentMethodProps> = ({
         async (options: PaymentInitializeOptions) => {
             return checkoutService.initializePayment({
                 ...options,
-                stripeupe: {
+                [method.gateway || method.id]: {
                     containerId,
+                    layout: {
+                        type: 'accordion',
+                        defaultCollapsed: selectedItemId !== methodSelector,
+                        radios: true,
+                        spacedAccordionItems: false,
+                        visibleAccordionItemsCount: 0,
+                    },
                     style: getStylesForOCSElement(containerId),
                     onError: onUnhandledError,
                     render: renderSubmitButton,
@@ -72,30 +80,55 @@ const StripeOCSPaymentMethod: FunctionComponent<PaymentMethodProps> = ({
                 },
             });
         },
-        [containerId, checkoutService, onUnhandledError, renderSubmitButton, onToggle],
+        [
+            containerId,
+            selectedItemId,
+            method,
+            methodSelector,
+            checkoutService,
+            onUnhandledError,
+            renderSubmitButton,
+            onToggle,
+        ],
     );
 
     const renderCheckoutThemeStylesForStripeOCS = () => {
         return (
-            <div style={{ display: 'none' }}>
-                <div id={`${containerId}--accordion-header`}>
-                    <div className="form-checklist-checkbox" />
-                    <div className="form-label optimizedCheckout-form-label" />
-                </div>
-                <div
-                    className="optimizedCheckout-form-input"
-                    id={`${containerId}--input`}
-                    placeholder="1111"
-                >
-                    <div className="form-field--error">
+            <>
+                <style>
+                    {`
+                        .custom-checklist-item#radio-${methodSelector} {
+                            border-bottom: none;
+                        }
+                        .custom-checklist-item#radio-${methodSelector}:last-of-type {
+                            margin-bottom: -1px;
+                        }
+                    `}
+                </style>
+
+                <div style={{ display: 'none' }}>
+                    <div id={`${containerId}--accordion-header`}>
+                        <div className="form-checklist-checkbox" />
+                        <div className="form-label optimizedCheckout-form-label" />
+                    </div>
+                    <div
+                        className="optimizedCheckout-form-input"
+                        id={`${containerId}--input`}
+                        placeholder="1111"
+                    >
+                        <div className="form-field--error">
+                            <div
+                                className="optimizedCheckout-form-label"
+                                id={`${containerId}--error`}
+                            />
+                        </div>
                         <div
                             className="optimizedCheckout-form-label"
-                            id={`${containerId}--error`}
+                            id={`${containerId}--label`}
                         />
                     </div>
-                    <div className="optimizedCheckout-form-label" id={`${containerId}--label`} />
                 </div>
-            </div>
+            </>
         );
     };